WIP: First/Last Hashed GBY:
Not working yet.
When FIRST is replaced with MIN(source_index), the results
are registered in the cache against the generated source_index
column, not the original source. They are thus not accessible
when finalizing the output for FIRST/NTH_ELEMENT.
The solution is to switch `flatten_single_pass_aggs()` to return
an object of a flattened_agg_input class, which has knowledge
of replaced inputs, etc. These can be replaced at the time the
output is registered in the cache.