Use `column_device_view.element()` instead of `.data()` for window range calculations.
This change is in preparation for supporting `STRING` order-by columns in range window functions.

In the staging step for executing window range queries, the boundaries of each row's window are
calculated. This involves subtracting/adding the `preceding`/`following` values from each
order-by column row, and then searching backwards/forwards for the boundary values.

The staging step has been using `column_device_view.data()` for accessing the order-by rows,
an acceptable approach for when the order-by columns are numeric (e.g. `INT32`).

This approach fails when the order-by column is a `STRING`, because `.data()` is not defined
for such columns. A better approach would be to use `.element()` to directly access the rows,
because it has special handling for `STRING`, among other types, while continuing to work
for numeric primitives.

In a followup to this change, support for `STRING` order-by columns will be added.

Signed-off-by: MithunR <mythrocks@gmail.com>